型定義専用のフォルダとファイルを作成して移動、それによるimportなどの変更
